lecture 5/who_do_you_look_like_bot_template/requirements.txt
torch
torchvision
facenet-pytorch
numpy
lmdb
faiss-cpu
pillow
aiogram
opencv-python-headless
kagglehub
scipy
